Mealmind/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
aiohttp
httpx
openai
langchain
celery
redis
python-multipart
jinja2
oauthlib
google-auth
google-auth-oauthlib
email-validator
